Tested the CRUD scripts
Tested and made sure the CRUD scripts worked, only with a token.
Can index without a token so viewers can see the articles.
Updated the schema to include the user ID and Overview of the articles
Deleted the old migration so I could include Overview and user ID.
Also fixed the relationship between user and articles.
So that authorization is required to create, update, or destroy.